feat(v2.1.1): cleanup SQLite — server/database.py purge legacy, server/main.py drop init_db startup, scripts isolés (refonte V2.1.2), README setup PG, V2.1 status: delivered (203/203 tests GREEN)
