+++ server/main.py
@@ -1,9 +1,10 @@
-from fastapi import FastAPI
-from fastapi.staticfiles import StaticFiles
-from fastapi.responses import FileResponse
 from pathlib import Path
-from server.database import init_db
-from server.routers import sleep, steps, heartrate, exercise
+
+from fastapi import FastAPI
+from fastapi.responses import FileResponse
+from fastapi.staticfiles import StaticFiles
+
+from server.routers import exercise, heartrate, sleep, steps
 
 app = FastAPI(title="SamsungHealth")
 app.include_router(sleep.router)
@@ -13,11 +14,6 @@ app.include_router(exercise.router)
 
 static_dir = Path(__file__).resolve().parent.parent / "static"
 app.mount("/static", StaticFiles(directory=str(static_dir)), name="static")
-
-
-@app.on_event("startup")
-def startup():
-    init_db()
 
 
 @app.get("/")
